Remove additional usage of service holder
