+++ components/rule/org.wso2.carbon.rule.backend/src/main/java/org/wso2/carbon/rule/backend/internal/RuleServiceValueHolder.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.wso2.carbon.rule.backend.internal;
+
+import org.wso2.carbon.registry.core.exceptions.RegistryException;
+import org.wso2.carbon.registry.core.service.RegistryService;
+
+public class RuleServiceValueHolder {
+    private RegistryService registryService;
+    private static RuleServiceValueHolder ruleServiceValueHolder = new RuleServiceValueHolder();
+
+
+    public static RuleServiceValueHolder getInstance() {
+        return ruleServiceValueHolder;
+    }
+
+    public RegistryService getRegistryService() throws RegistryException {
+        return registryService;
+    }
+
+    public void setRegistryService(RegistryService registryService) {
+        this.registryService = registryService;
+    }
+
+    public void unsetRegistryService() {
+        this.registryService = null;
+    }
+
+}
