+++ pom.xml
@@ -588,7 +588,7 @@
         <carbon.kernel.registry.imp.pkg.version>[1.0.1, 2.0.0)</carbon.kernel.registry.imp.pkg.version>
 
         <!-- Carbon commons version -->
-        <carbon.commons.version>4.7.0</carbon.commons.version>
+        <carbon.commons.version>4.7.1</carbon.commons.version>
         <carbon.commons.imp.pkg.version>[4.7.0, 4.8.0)</carbon.commons.imp.pkg.version>
 
         <!-- Carbon Identity version -->
@@ -596,7 +596,7 @@
         <!--<carbon.identity.imp.pkg.version>[5.0.0, 6.0.0)</carbon.identity.imp.pkg.version>-->
         
         <!-- Carbon Deployment version -->
-        <carbon.deployment.version>4.9.0</carbon.deployment.version>
+        <carbon.deployment.version>4.9.4</carbon.deployment.version>
         <carbon.deployment.imp.pkg.version>[4.9.0, 5.0.0)</carbon.deployment.imp.pkg.version>
 
         <!-- Carbon registry version -->
@@ -626,7 +626,7 @@
         <version.javax.servlet>3.0.0.v201112011016</version.javax.servlet>
 
         <!-- Rampart Version -->
-        <rampart.wso2.version>1.6.1-wso2v16</rampart.wso2.version>
+        <rampart.wso2.version>1.6.1-wso2v36</rampart.wso2.version>
 
         <libthrift.wso2.version>0.8.0.wso2v1</libthrift.wso2.version>
